+++ Runtime/Scripts/Utilities/Recycling/IRecyclable.cs
@@ -1,0 +1,13 @@
+﻿using UnityEngine;
+
+namespace Utilities.Recycling
+{
+    public interface IRecyclable
+    {
+        GameObject gameObject { get; }
+        Transform transform { get; }
+
+        bool IsRecycled { get; set; }
+        void OnRecycled();
+    }
+}