+++ CHANGELOG.md
@@ -14,10 +14,24 @@ and this project adheres to [Semantic Versioning](http://semver.org/).
 - 
 
 ### Fixed
+-
+
+## [0.0.2] - 2024-12-29
+
+### Added
+- `IRecyclable.cs` as interface to include on classes that need `GameObject` recycled
+- `Recycler.cs` as `static class` that can attempt to recycle & retrieve `GameObject`
+  - This will also auto-generate a Parent Transform when requested.
+- Recycling Sample into Package
+  - This includes showcasing the difference of recycle via `Enum` or `IRecyclable`
+
+### Changed
 - 
 
-## [0.0.1] - 2024-12-15
+### Fixed
+-
 
+## [0.0.1] - 2024-12-15
 
 ### Added
 - All documentation of included scripts & Samples